+++ dto/token_issue_response.go
@@ -1,5 +1,5 @@
 //
-// Copyright (C) 2019 Authlete, Inc.
+// Copyright (C) 2019-2020 Authlete, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -60,4 +60,7 @@ type TokenIssueResponse struct {
 
 	//
 	JwtAccessToken string `json:"jwtAccessToken,omitempty"`
+
+	//
+	AccessTokenResources []string `json:"accessTokenResources,omitempty"`
 }