+++ CHANGES.md
@@ -77,6 +77,9 @@ CHANGES
 - `TokenFailReason` type
     * Added `TokenFailReason_INVALID_TARGET`.
 
+- `TokenIssueResponse` struct
+    * Added `AccessTokenResources` field.
+
 - New components
     * `PushedAuthReqAction` type and some of its constants
     * `PushedAuthReqRequest` struct